update module layout for better code organization
