+++ src/test/http.rs
@@ -1,0 +1,74 @@
+use hyper::Method;
+use serde::{Deserialize, Serialize};
+use std::cell::Cell;
+use std::hash::{Hash, Hasher};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct HttpHeader {
+    pub header: String,
+    pub value: String,
+
+    #[serde(skip_serializing, skip_deserializing)]
+    pub matches_variable: Cell<bool>,
+}
+
+impl Hash for HttpHeader {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.header.hash(state);
+        self.value.hash(state);
+    }
+
+    fn hash_slice<H: Hasher>(data: &[Self], state: &mut H)
+    where
+        Self: Sized,
+    {
+        for piece in data {
+            piece.hash(state);
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct HttpParameter {
+    pub param: String,
+    pub value: String,
+
+    #[serde(skip_serializing, skip_deserializing)]
+    pub matches_variable: Cell<bool>,
+}
+
+impl Hash for HttpParameter {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.param.hash(state);
+        self.value.hash(state);
+    }
+
+    fn hash_slice<H: Hasher>(data: &[Self], state: &mut H)
+    where
+        Self: Sized,
+    {
+        for piece in data {
+            piece.hash(state);
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Hash)]
+pub enum HttpVerb {
+    Get,
+    Post,
+    Put,
+    Patch,
+    Undefined,
+}
+
+impl HttpVerb {
+    pub fn as_method(&self) -> Method {
+        match &self {
+            HttpVerb::Post => Method::POST,
+            HttpVerb::Patch => Method::PATCH,
+            HttpVerb::Put => Method::PUT,
+            _ => Method::GET,
+        }
+    }
+}
