+++ src/test/mod.rs
@@ -1,0 +1,5 @@
+pub mod definition;
+pub mod file;
+pub mod http;
+pub mod templates;
+pub mod validator;
